# swagger.yaml
openapi: 3.0.3
info:
  title: IDM App API
  description: |
    Identity and Access Management (IDM) Application API
    
    This API provides comprehensive user, group, organization, and role management capabilities
    with Auth0 integration for secure authentication and authorization.
    
    ## Features
    - User Management (CRUD operations)
    - Group Management (CRUD operations)
    - Organization Management (CRUD operations)
    - Role Management (CRUD operations)
    - Group Member Management (User-Group relationships)
    - Organization Member Management (User/Group/Role-Organization relationships)
    - Role Member Management (User/Group-Role relationships)
    - Auth0 JWT Authentication
    - Health Monitoring
    
    ## Authentication
    All protected endpoints require a valid JWT token from Auth0. Include the token in the Authorization header:
    ```
    Authorization: Bearer <your-jwt-token>
    ```
    
    ## Getting Started
    1. Configure your Auth0 credentials in the `.env` file
    2. Obtain a JWT token using the Auth0 client credentials flow
    3. Use the token to access protected endpoints
    
    ## Note on Missing Endpoints
    The following endpoints exist in the codebase but are not currently wired up in the routes:
    - Organization Member Management (`/api/v1/orgmembers`) - Add/remove users, groups, or roles from organizations
    - Role Member Management (`/api/v1/rolemembers`) - Add/remove users or groups from roles
    
...
    
...
servers:
...
security:
  - BearerAuth: []
paths:
...
  /api/v1/groups:
    get:
      tags:
        - Groups
      summary: Get all groups
      description: Retrieve a list of all groups in the system
      operationId: getAllGroups
      security:
        - BearerAuth: []
      responses:
        '200':
          description: List of groups retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/GroupResponse'
        '401':
          description: Unauthorized - Invalid or missing token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    post:
      tags:
        - Groups
      summary: Create a new group
      description: Create a new group in the system
      operationId: createGroup
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GroupCreateRequest'
      responses:
        '201':
          description: Group created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GroupResponse'
        '400':
          description: Bad request - Invalid input data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized - Invalid or missing token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Conflict - Group with this name already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/groups/{id}:
    get:
      tags:
        - Groups
      summary: Get group by ID
      description: Retrieve a specific group by its ID
      operationId: getGroupById
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Group ID (UUID)
          schema:
            type: string
            format: uuid
            example: "123e4567-e89b-12d3-a456-426614174000"
      responses:
        '200':
          description: Group retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GroupResponse'
        '401':
          description: Unauthorized - Invalid or missing token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Group not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    put:
      tags:
        - Groups
      summary: Update group
      description: Update an existing group's information
      operationId: updateGroup
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Group ID (UUID)
          schema:
            type: string
            format: uuid
            example: "123e4567-e89b-12d3-a456-426614174000"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GroupUpdateRequest'
      responses:
        '200':
          description: Group updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GroupResponse'
        '400':
          description: Bad request - Invalid input data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized - Invalid or missing token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Group not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Conflict - Group name already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    delete:
      tags:
        - Groups
      summary: Delete group
      description: Delete a group from the system
      operationId: deleteGroup
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Group ID (UUID)
          schema:
            type: string
            format: uuid
            example: "123e4567-e89b-12d3-a456-426614174000"
      responses:
        '200':
          description: Group deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Group deleted successfully"
        '401':
          description: Unauthorized - Invalid or missing token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Group not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/v1/groupmembers/group/{groupId}:
    get:
      tags:
        - Group Members
      summary: Get members by group ID
      description: Retrieve all members of a specific group
      operationId: getMembersByGroupId
      security:
        - BearerAuth: []
      parameters:
        - name: groupId
          in: path
          required: true
          description: Group ID (UUID)
          schema:
            type: string
            format: uuid
            example: "123e4567-e89b-12d3-a456-426614174000"
      responses:
        '200':
          description: Group members retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MemberResponse'
        '400':
          description: Bad request - Invalid group ID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized - Invalid or missing token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden - Insufficient permissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/groupmembers/user/{userId}:
    get:
      tags:
        - Group Members
      summary: Get memberships by user ID
      description: Retrieve all group memberships for a specific user
      operationId: getMembersByUserId
      security:
        - BearerAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          description: User ID (UUID)
          schema:
            type: string
            format: uuid
            example: "123e4567-e89b-12d3-a456-426614174000"
      responses:
        '200':
          description: User memberships retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MemberResponse'
        '400':
          description: Bad request - Invalid user ID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized - Invalid or missing token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden - Insufficient permissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT token obtained from Auth0. Include the token in the Authorization header.
        Example: `Authorization: Bearer eyJhbGciOiJSUzI1NiIsInR5cCI6IkpXVCJ9...`
  schemas:
...
    GroupCreateRequest:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          description: Unique group name
          example: "developers"
        displayName:
          type: string
          description: Display name for the group
          example: "Development Team"
        description:
          type: string
          description: Group description
          example: "Software development team"
    GroupUpdateRequest:
      type: object
      properties:
        name:
          type: string
          description: Unique group name
          example: "senior-developers"
        displayName:
          type: string
          description: Display name for the group
          example: "Senior Development Team"
        description:
          type: string
          description: Group description
          example: "Senior software development team"
        isActive:
          type: boolean
          description: Whether the group is active
          example: true
    GroupResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: Group ID
          example: "123e4567-e89b-12d3-a456-426614174000"
        name:
          type: string
          description: Group name
          example: "developers"
        displayName:
          type: string
          description: Display name
          example: "Development Team"
        description:
          type: string
          description: Group description
          example: "Software development team"
        isActive:
          type: boolean
          description: Active status
          example: true
        createdAt:
          type: string
          format: date-time
          description: Creation timestamp
          example: "2025-06-22T14:00:00Z"
        updatedAt:
          type: string
          format: date-time
          description: Last update timestamp
          example: "2025-06-22T14:00:00Z"
...
    MemberResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: Member ID
          example: "123e4567-e89b-12d3-a456-426614174000"
        groupId:
          type: string
          format: uuid
          description: Group ID
          example: "123e4567-e89b-12d3-a456-426614174000"
        userId:
          type: string
          format: uuid
          description: User ID
          example: "123e4567-e89b-12d3-a456-426614174000"
        createdAt:
          type: string
          format: date-time
          description: Creation timestamp
          example: "2025-06-22T14:00:00Z"
        updatedAt:
          type: string
          format: date-time
          description: Last update timestamp
          example: "2025-06-22T14:00:00Z"
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Error message
          example: "User not found"
        code:
          type: string
          description: Error code
          example: "USER_NOT_FOUND"
        details:
          type: object
          description: Additional error details
          additionalProperties: true
...